refactor: add seed deriviation function for decoupling key gen logic with it
