+++ crates/common/account_manager/Cargo.toml
@@ -20,8 +20,4 @@ rand_chacha.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
-tracing.workspace = true
 uuid = { version = "1.0", features = ["v4", "serde"] }
-
-# ream dependencies
-ream-post-quantum-crypto.workspace = true
